Add extracted view components
