+++ MisMangasACA/Views/HomeView.swift
@@ -247,280 +247,3 @@ struct HomeView: View {
     }
 }
 
-// Fila sencilla para cada manga
-struct MangaRowView: View {
-    let manga: MangaDTO
-    
-    var body: some View {
-        let cleanedURL = manga.mainPicture
-            .flatMap { raw -> URL? in
-                let trimmed = raw.replacingOccurrences(of: "\"", with: "")
-                return URL(string: trimmed)
-            }
-        
-        ViewThatFits {
-            // Horizontal layout
-            HStack(alignment: .top, spacing: 12) {
-                coverImage(from: cleanedURL)
-                mangaDetails
-                Spacer()
-            }
-            
-            // Vertical layout
-            VStack(alignment: .leading, spacing: 8) {
-                coverImage(from: cleanedURL)
-                mangaDetails
-            }
-        }
-        .padding()
-        .background(.ultraThinMaterial)
-        .cornerRadius(12)
-        .shadow(radius: 2)
-    }
-    
-    @ViewBuilder
-    private var mangaDetails: some View {
-        VStack(alignment: .leading, spacing: 4) {
-            ViewThatFits {
-                // Pantallas grandes: 1 línea truncada
-                Text(manga.title)
-                    .font(.title3.bold())
-                    .lineLimit(1)
-                    .truncationMode(.tail)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-
-                // Pantallas chicas: hasta 2 líneas
-                Text(manga.title)
-                    .font(.title3.bold())
-                    .lineLimit(2)
-                    .multilineTextAlignment(.leading)
-                    .frame(maxWidth: .infinity, alignment: .leading)
-            }
-
-            if let score = manga.score {
-                Text("⭐️ \(String(format: "%.1f", score))")
-                    .font(.subheadline)
-                    .foregroundColor(.secondary)
-            }
-
-            if let status = manga.status {
-                Text(status.replacingOccurrences(of: "_", with: " ").capitalized)
-                    .font(.footnote)
-                    .foregroundColor(.secondary)
-            }
-
-            if let startDate = manga.startDate {
-                let year = Calendar.current.component(.year, from: startDate)
-                Text("📅 \(year)")
-                    .font(.footnote)
-                    .foregroundColor(.secondary)
-            }
-
-            if !manga.genres.isEmpty {
-                HStack(spacing: 4) {
-                    ForEach(manga.genres.prefix(2), id: \.id) { genre in
-                        Text(genre.genre)
-                            .font(.caption2)
-                            .padding(.horizontal, 6)
-                            .padding(.vertical, 2)
-                            .background(Color.accentColor.opacity(0.1))
-                            .cornerRadius(4)
-                    }
-                }
-            }
-        }
-    }
-    
-    private func coverImage(from url: URL?) -> some View {
-        AsyncImage(url: url) { phase in
-            switch phase {
-            case .empty:
-                Color.gray
-            case .success(let img):
-                img.resizable().scaledToFill()
-            case .failure:
-                Color.red
-            @unknown default:
-                EmptyView()
-            }
-        }
-        .frame(width: 80, height: 120)
-        .clipped()
-        .cornerRadius(8)
-        .shadow(radius: 2)
-        .accessibilityLabel(Text("Portada de \(manga.title)"))
-    }
-}
-
-struct MainTabView: View {
-    var body: some View {
-        TabView {
-            HomeView()
-                .tabItem {
-                    Label("Mangas", systemImage: "books.vertical")
-                }
-            AuthorsView()
-                .tabItem {
-                    Label("Autores", systemImage: "person.3.sequence.fill")
-                }
-            CollectionView()
-                .tabItem {
-                    Label("Colección", systemImage: "books.vertical")
-                }
-        }
-    }
-}
-
-// MARK: – Panel de filtros avanzado
-@available(iOS 17.0, macOS 14.0, *)
-struct FilterInspector: View {
-    @ObservedObject var vm: HomeViewModel
-    @Environment(\.dismiss) private var dismiss     // Para cerrar el inspector/sheet
-
-    // Wrapper que da identidad estable a cada opción de catálogo
-    private struct OptionRow: Identifiable, Hashable {
-        let id = UUID()
-        let label: String
-    }
-
-    // Colecciones con identidad estable (se rellenan en onAppear)
-    @State private var demoRows: [OptionRow] = []
-    @State private var themeRows: [OptionRow] = []
-
-    var body: some View {
-        NavigationStack {
-            Form {
-                // Campo de texto para título
-                Section("Título") {
-                    TextField(
-                        "",
-                        text: $vm.filterSearchText,
-                        prompt: Text(vm.filterContains ? "Contiene…" : "Empieza por…")
-                    )
-                    .textInputAutocapitalization(.never)
-                    .disableAutocorrection(true)
-                }
-
-                // Segmentado Contiene / Empieza por
-                Section("Coincidencia") {
-                    Picker("", selection: $vm.filterContains) {
-                        Text("Contiene").tag(true)
-                        Text("Empieza por").tag(false)
-                    }
-                    .pickerStyle(.segmented)
-                }
-
-                // Género (selección única)
-                if !vm.genres.isEmpty {
-                    Section("Género") {
-                        Picker("Seleccionar género", selection: $vm.selectedGenre) {
-                            Text("Todos").tag(String?.none)
-                            ForEach(vm.genres, id: \.self) { genre in
-                                Text(genre).tag(String?.some(genre))
-                            }
-                        }
-                        .labelsHidden()          // Evita la fila vacía en el picker
-                        .pickerStyle(.inline)
-                    }
-                }
-
-                // Demografías (multi‑selección)
-                if !vm.demographics.isEmpty {
-                    Section("Demografía") {
-                        ForEach(demoRows) { row in
-                            Toggle(row.label, isOn: Binding(
-                                get: { vm.selectedDemographies.contains(row.label) },
-                                set: { isOn in
-                                    if isOn {
-                                        vm.selectedDemographies.insert(row.label)
-                                    } else {
-                                        vm.selectedDemographies.remove(row.label)
-                                    }
-                                }))
-                        }
-                    }
-                }
-
-                // Temáticas (multi‑selección)
-                if !vm.themes.isEmpty {
-                    Section("Temática") {
-                        ForEach(themeRows) { row in
-                            Toggle(row.label, isOn: Binding(
-                                get: { vm.selectedThemes.contains(row.label) },
-                                set: { isOn in
-                                    if isOn {
-                                        vm.selectedThemes.insert(row.label)
-                                    } else {
-                                        vm.selectedThemes.remove(row.label)
-                                    }
-                                }))
-                        }
-                    }
-                }
-            }
-            .onAppear {
-                if demoRows.isEmpty {
-                    demoRows  = vm.demographics.map { OptionRow(label: $0) }
-                    themeRows = vm.themes.map        { OptionRow(label: $0) }
-                }
-            }
-            .navigationTitle("Filtros")
-            .toolbar {
-                // Botón Limpiar
-                ToolbarItem(placement: .cancellationAction) {
-                    Button("Limpiar") {
-                        dismiss()   // Cierra primero
-                        Task {
-                            vm.resetFilters()
-                            await vm.loadPage(1, forceReload: true)
-                        }
-                    }
-                    .foregroundColor(.accentColor)
-                }
-                // Botón Aplicar
-                ToolbarItem(placement: .confirmationAction) {
-                    Button("Aplicar") {
-                        dismiss()   // Cierra primero
-                        Task {
-                            await vm.applyFilters()
-                        }
-                    }
-                    .foregroundColor(.accentColor)
-                }
-            }
-            .tint(.accentColor)   // Fuerza botones azules en cualquier estado del sheet
-        }
-    }
-}
-
-// MARK: – Celda compacta para grid
-/*
- struct MangaCoverCell: View {
-    let manga: MangaDTO
-    
-    var body: some View {
-        VStack(spacing: 8) {
-            AsyncImage(url: manga.mainPicture.flatMap {
-                URL(string: $0.replacingOccurrences(of: "\"", with: ""))
-            }) { phase in
-                switch phase {
-                case .empty: Color.gray.opacity(0.3)
-                case .success(let img): img.resizable().scaledToFill()
-                case .failure: Color.red
-                @unknown default: Color.gray
-                }
-            }
-            .frame(width: 140, height: 200)
-            .clipped()
-            .cornerRadius(8)
-            .shadow(radius: 1)
-            
-            Text(manga.title)
-                .font(.caption)
-                .multilineTextAlignment(.center)
-                .lineLimit(2)
-                .frame(width: 140)
-        }
-    }
-}
-*/
